Update and rename TestAssignment2.java to TestQueenSolution.java
